requirements.txt
aiomysql==0.2.0
APScheduler==3.11.0
pydantic==2.10.6
python-dotenv==1.0.1
pytz==2025.1
Telethon==1.37.0
asyncio